# Speed up `new_tibble()`