--- a/R/new.R
+++ b/R/new.R
@@ -76,7 +76,9 @@ new_tibble <- function(x, ..., nrow, class = NULL, subclass = NULL) {
     cnd_signal(error_names_must_be_non_null())
   }
 
-  set_tibble_subclass(x, nrow, class)
+  attr(x, "row.names") <- .set_row_names(nrow)
+  class(x) <- c(class[!class %in% tibble_class], tibble_class)
+  x
 }
 
 #' @description
@@ -134,13 +136,6 @@ update_tibble_attrs <- function(x, ...) {
 }
 
 tibble_class <- c("tbl_df", "tbl", "data.frame")
-
-# Two dedicated functions for faster creation
-set_tibble_subclass <- function(x, nrow, subclass) {
-  attr(x, "row.names") <- .set_row_names(nrow)
-  class(x) <- c(class[!class %in% tibble_class], tibble_class)
-  x
-}
 
 # Errors ------------------------------------------------------------------